requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
blinker==1.8.2
boto3==1.35.38
botocore==1.35.38
Brotli==1.1.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.0
ffmpy==0.4.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
gradio==5.0.1
gradio_client
gunicorn
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
mutagen==1.47.0
numpy==2.1.1
openai==1.51.0
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pysrt==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
ruff==0.6.9
s3transfer==0.10.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
websockets
Werkzeug==3.0.4
whisper==1.1.10
youtube-dl==2021.12.17
yt-dlp
openai-whisper